Removes SK2 Purchasing logic, adds caching policy
- StoreKit 2 was too buggy to be able to use :(
- Updates AppSessionLogic with Bryan's code.
- Stores the active subscription status to disk.
- Keep loading indicator displaying for users who are manually controlling dismissal of the paywall via the automaticallyDismiss option
- Updated caching policy for requests to our servers such that it uses the last-modified header.
- Updates changelog
