+++ Sources/SuperwallKit/StoreKit/Abstractions/StoreProduct/SubscriptionPeriod.swift
@@ -82,16 +82,6 @@ public final class SubscriptionPeriod: NSObject {
       .normalized()
   }
 
-  @available(iOS 15.0, tvOS 15.0, watchOS 8, *)
-  static func from(sk2SubscriptionPeriod: StoreKit.Product.SubscriptionPeriod) -> SubscriptionPeriod? {
-    guard let unit = SubscriptionPeriod.Unit.from(sk2PeriodUnit: sk2SubscriptionPeriod.unit) else {
-      return nil
-    }
-
-    return .init(value: sk2SubscriptionPeriod.value, unit: unit)
-      .normalized()
-  }
-
   /// This function simplifies large numbers of days into months and large numbers
   /// of months into years if there are no leftover units after the conversion.
   ///
@@ -191,17 +181,6 @@ extension SubscriptionPeriod {
 fileprivate extension SubscriptionPeriod.Unit {
   static func from(sk1PeriodUnit: SK1Product.PeriodUnit) -> Self? {
     switch sk1PeriodUnit {
-    case .day: return .day
-    case .week: return .week
-    case .month: return .month
-    case .year: return .year
-    @unknown default: return nil
-    }
-  }
-
-  @available(iOS 15.0, tvOS 15.0, watchOS 8, *)
-  static func from(sk2PeriodUnit: StoreKit.Product.SubscriptionPeriod.Unit) -> Self? {
-    switch sk2PeriodUnit {
     case .day: return .day
     case .week: return .week
     case .month: return .month
