+++ Sources/SuperwallKit/StoreKit/Coordinator/StoreKitCoordinator.swift
@@ -9,6 +9,16 @@ import Foundation
 
 /// Coordinates: The purchasing, restoring and retrieving of products; the checking
 /// of transactions; and the determining of the user's subscription status.
+///
+/// We can't use StoreKit 2 because:
+/// 1. If there are unfinished StoreKit transaction updates on app open, only the first
+///    listener is given the updates. This means that our listener can interfere with
+///    the clients listener (if they're using SK2) and their transactions may not even show up!
+/// 2. If the developer used an SK1 listener, their listener could finish the txns before
+///    ours. Which means that we will never see the unfinished transactions or we'll see only
+///    a few of them.
+/// 3. The Transaction.updates listener is not reliably called for transactions made outside
+///   the app. Therefore the transactions aren't always finished.
 struct StoreKitCoordinator {
   /// Purchases the product.
   let productPurchaser: ProductPurchaser
@@ -32,52 +42,18 @@ struct StoreKitCoordinator {
     purchasingDelegateAdapter: SuperwallPurchasingDelegateAdapter,
     storeKitManager: StoreKitManager,
     finishTransactions: Bool,
-    sessionEventsManager: SessionEventsManager,
-    factory: StoreTransactionFactory
+    factory: StoreTransactionFactory & ProductPurchaserFactory
   ) {
     let hasDelegate = purchasingDelegateAdapter.hasDelegate
 
     if hasDelegate {
-      // If delegate exists, finishTransactions will always be false and SK1 must be used
-      // for both Superwall and in the developer's purchasing delegate. No mixing and matching
-      // of SK1 and SK2.
-      // In addition, after testing, it turns out we can't use SK2 here because:
-      // 1. If there are unfinished StoreKit transaction updates on app open, only the first
-      //    listener is given the updates. This means that our listener can interfere with
-      //    the clients listener (if they're using SK2) and their transactions may not even show up!
-      // 2. If they use an SK1 listener, their listener could finish the txns before
-      //    ours. Which means that we will never see the unfinished transactions or we'll see only
-      //    a few of them.
       self.productPurchaser = purchasingDelegateAdapter
       self.productFetcher = ProductsFetcherSK1()
-      self.txnChecker = ProductPurchaserSK1(
-        storeKitManager: storeKitManager,
-        sessionEventsManager: sessionEventsManager,
-        factory: factory
-      )
+      self.txnChecker = factory.makeSK1ProductPurchaser()
       self.txnRestorer = purchasingDelegateAdapter
       self.subscriptionStatusHandler = purchasingDelegateAdapter
-    } else if #available(iOS 15.0, tvOS 15.0, watchOS 8.0, *),
-      finishTransactions {
-      // Use SK2 when Superwall handles everything on iOS 15+.
-      let purchaser = ProductPurchaserSK2(
-        storeKitManager: storeKitManager,
-        sessionEventsManager: sessionEventsManager,
-        factory: factory
-      )
-      self.productPurchaser = purchaser
-      self.productFetcher = ProductsFetcherSK2()
-      self.txnChecker = purchaser
-      self.txnRestorer = purchaser
-      self.subscriptionStatusHandler = storeKitManager
     } else {
-      // Use SK1 on <iOS 15 or when the user is finishing
-      // transactions themselves.
-      let purchaser = ProductPurchaserSK1(
-        storeKitManager: storeKitManager,
-        sessionEventsManager: sessionEventsManager,
-        factory: factory
-      )
+      let purchaser = factory.makeSK1ProductPurchaser()
       self.productPurchaser = purchaser
       self.productFetcher = ProductsFetcherSK1()
       self.txnChecker = purchaser
